Remake some const to compeont
